[bumble-tech#59] Changed ParentNodeView to be interface
