+++ core/src/main/java/com/bumble/appyx/core/node/AbstractParentNodeView.kt
@@ -1,0 +1,19 @@
+package com.bumble.appyx.core.node
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+abstract class AbstractParentNodeView<Routing : Any> : ParentNodeView<Routing> {
+
+    final override lateinit var node: ParentNode<Routing>
+        private set
+
+    override fun init(node: ParentNode<Routing>) {
+        this.node = node
+    }
+
+    @Composable
+    override fun View(modifier: Modifier) {
+        node.NodeView(modifier = modifier)
+    }
+}
