+++ core/src/main/java/com/bumble/appyx/core/node/ParentNodeView.kt
@@ -4,17 +4,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.bumble.appyx.core.plugin.NodeAware
 
-abstract class ParentNodeView<Routing : Any> : ParentNodeViewBase<Routing> {
-
-    final override lateinit var node: ParentNode<Routing>
-        private set
-
-    override fun init(node: ParentNode<Routing>) {
-        this.node = node
-    }
-
-    @Composable
-    override fun View(modifier: Modifier) {
-        node.NodeView(modifier = modifier)
-    }
+interface ParentNodeView<Routing : Any> : NodeView, NodeAware<ParentNode<Routing>> {
+    @Composable fun ParentNode<Routing>.NodeView(modifier: Modifier)
 }
