--- core/src/main/java/com/bumble/appyx/core/node/ParentNodeViewBase.kt
@@ -1,9 +1,0 @@
-package com.bumble.appyx.core.node
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import com.bumble.appyx.core.plugin.NodeAware
-
-interface ParentNodeViewBase<Routing : Any> : NodeView, NodeAware<ParentNode<Routing>> {
-    @Composable fun ParentNode<Routing>.NodeView(modifier: Modifier)
-}
