+++ sandbox/src/main/java/com/bumble/appyx/sandbox/client/mvicoreexample/MviCoreExampleView.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.bumble.appyx.core.composable.Children
+import com.bumble.appyx.core.node.AbstractParentNodeView
 import com.bumble.appyx.core.node.ParentNode
-import com.bumble.appyx.core.node.ParentNodeView
 import com.bumble.appyx.core.routing.RoutingSource
 import com.bumble.appyx.routingsource.backstack.BackStack
 import com.bumble.appyx.routingsource.backstack.transitionhandler.rememberBackstackSlider
@@ -47,7 +47,7 @@ class MviCoreExampleViewImpl(
     private val title: String = "Title",
     private val backStack: RoutingSource<Routing, BackStack.TransitionState>,
     private val events: PublishRelay<Event> = PublishRelay.create()
-) : ParentNodeView<Routing>(),
+) : AbstractParentNodeView<Routing>(),
     MviCoreExampleView,
     ObservableSource<Event> by events {
